task_9/requirements.txt
langchain
langchain-openai
langchain-community
langgraph
chromadb
openai
python-dotenv

